Move functions from Mix.Tasks.Compile to Mix.Task.Compiler
